Weighted Reaction System 2.0 (#91)
* Add Reaction rate attribute

* Change weight attributes to float

* Change weight db column types

* Add rate attribute to Reacter::reactTo method

* Add ability to rewrite reaction rate if differs from persisted one

* Add getRate method to Reaction

* Add upgrade v7 to v8 Artisan command

* Extract default ReactionType mass to constant

* Add $rate property to reactTo method of Reacter facade

* Add $rate to isReactedBy & isNotReactedBy methods

* Add $rate to hasReactedTo & hasNotReactedTo methods

* Define minimum & maximum Reaction rate values

* Add RateOutOfRange exception & min + max rate constants

+++ contracts/Reactant/ReactionTotal/Models/ReactionTotal.php
@@ -21,13 +21,13 @@ interface ReactionTotal
 
     public function getCount(): int;
 
-    public function getWeight(): int;
+    public function getWeight(): float;
 
     public function incrementCount(int $amount): void;
 
     public function decrementCount(int $amount): void;
 
-    public function incrementWeight(int $amount): void;
+    public function incrementWeight(float $amount): void;
 
-    public function decrementWeight(int $amount): void;
+    public function decrementWeight(float $amount): void;
 }
