+++ database/migrations/2018_07_25_001000_create_love_reactant_reaction_totals_table.php
@@ -28,7 +28,7 @@ final class CreateLoveReactantReactionTotalsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('reactant_id');
             $table->unsignedBigInteger('count');
-            $table->bigInteger('weight');
+            $table->decimal('weight', 13, 2);
             $table->timestamps();
 
             $table
